Use a minimal docker image for autopkgtests
Since autopkgtest is using the null virtualization server (aka, the
existing environment), we don't want to re-use the build environment to
run the tests.  Not only does it have extra packages installed, but it
can break the "needs-recommends" Restriction.

Case in point, devscripts uses this Restriction, but since devscripts is
installed in the build environment (using --no-install-recommends) its
autopkgtests fail due to missing packages.